npm install --save-dev @babel/plugin-proposal-private-property-in-object
Removed warning: One of your dependencies, babel-preset-react-app, is importing the"@babel/plugin-proposal-private-property-in-object" package withoutdeclaring it in its dependencies. This is currently working because"@babel/plugin-proposal-private-property-in-object" is already in yournode_modules folder for unrelated reasons, but it may break at any time.babel-preset-react-app is part of the create-react-app project, whichis not maintianed anymore. It is thus unlikely that this bug willever be fixed. Add "@babel/plugin-proposal-private-property-in-object" toyour devDependencies to work around this error. This will make this messagego away.